requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=0.15.0=pyhd3eb1b0_0
alabaster=0.7.12=py36_0
appdirs=1.4.4=pyhd3eb1b0_0
argh=0.26.2=py36_0
argon2-cffi=19.2.0=pypi_0
asgiref=3.4.1=pypi_0
asn1crypto=1.3.0=py36_0
astor=0.8.1=py36haa95532_0
astroid=2.3.3=py36_0
atomicwrites=1.3.0=py36_1
attrs=19.3.0=py_0
autopep8=1.4.4=py_0
babel=2.8.0=py_0
backcall=0.1.0=py36_0
bcrypt=3.1.7=py36he774522_0
blas=1.0=mkl
bleach=3.1.4=py_0
blinker=1.4=py36haa95532_0
brotli=1.0.9=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
certifi=2021.5.30=py36haa95532_0
cffi=1.14.0=py36h7a1dbc1_0
chardet=3.0.4=py36_1003
click=8.0.4=pypi_0
click-plugins=1.1.1=pyhd3eb1b0_0
cligj=0.7.2=pypi_0
cloudpickle=1.3.0=py_0
colorama=0.4.3=py_0
conda-pack=0.4.0=pypi_0
construct=2.10.56=pypi_0
coverage=5.5=py36h2bbff1b_2
cryptography=2.8=py36h7a1dbc1_0
curl=7.82.0=h2bbff1b_0
cx_oracle=7.2.3=py36h62dcd97_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.24=py36hd77b12b_0
dataclasses=0.8=pyh4f3eec9_6
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
deprecated=1.2.12=pyhd3eb1b0_0
descartes=1.1.0=pyhd3eb1b0_4
diff-match-patch=20181111=py_0
django=3.2.16=pypi_0
django-cors-headers=3.10.1=pypi_0
django-python3-ldap=0.15.4=pypi_0
djangorestframework=3.14.0=pypi_0
docker-py=4.0.2=py36_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
docutils=0.16=py36_0
entrypoints=0.3=py36_0
et_xmlfile=1.1.0=py36haa95532_0
expat=2.4.4=h6c2663c_0
fiona=1.8.22=pypi_0
flake8=3.7.9=py36_0
flask=1.1.2=py_0
flask-compress=1.13=pypi_0
flask-enterprise=1.0=pypi_0
freetype=2.10.4=hd328e21_0
freexl=1.0.6=h2bbff1b_0
future=0.18.2=py36_0
gast=0.5.3=pyhd3eb1b0_0
gdal=2.3.3=py36hdf43c64_0
geopandas=0.9.0=pypi_0
geos=3.7.1=h33f27b4_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.3=py_1
greenlet=2.0.2=pypi_0
grpcio=1.35.0=pypi_0
h5py=2.10.0=py36h5e291fa_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=py36_0
imagesize=1.2.0=py_0
importlib-metadata=2.0.0=pypi_0
importlib_metadata=1.5.0=py36_0
intel-openmp=2020.0=166
intervaltree=3.0.2=py_0
ipykernel=5.1.4=py36h39e3cac_0
ipython=7.13.0=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.4.2=py36_0
isort=4.3.21=py36_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.15.2=py36_0
jinja2=2.11.1=pypi_0
jira=2.0.0=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py36_0
jupyter_client=6.1.2=py_0
jupyter_core=4.6.3=py36_0
kealib=1.4.7=h07cbb95_6
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
keyring=21.1.1=pypi_0
kiwisolver=1.2.0=pypi_0
krb5=1.16.4=hc04afaa_0
lazy-object-proxy=1.4.3=py36he774522_0
ldap3=2.9.1=pyhd3eb1b0_0
lerc=3.0=hd77b12b_0
libboost=1.73.0=h6c2663c_11
libcurl=7.82.0=h86230a5_0
libdeflate=1.8=h2bbff1b_5
libgdal=2.3.3=h10f50ba_0
libiconv=1.15=h1df5818_7
libkml=1.3.0=hd124aa8_5
libnetcdf=4.6.1=h411e497_2
libpng=1.6.37=h2a8f88b_0
libpq=11.2=h3235a2c_0
libprotobuf=3.17.2=h23ce68f_1
libsodium=1.0.16=h9d3ae62_0
libspatialindex=1.9.3=h33f27b4_0
libspatialite=4.3.0a=hc36aec2_19
libssh2=1.10.0=hcd4344a_0
libtiff=4.5.0=h8a3f274_0
libxml2=2.9.9=h464c3ec_0
libxslt=1.1.33=h579f668_0
lxml=4.5.0=py36h1350720_0
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mapproxy=1.14.0=pypi_0
markdown=3.3.4=py36haa95532_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.2.2=pypi_0
matplotlib-base=3.3.4=py36h49ac443_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36he774522_0
mkl=2020.0=166
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.0.15=py36h14836fe_0
mkl_random=1.1.0=py36h675688f_0
more-itertools=8.2.0=py_0
msys2-conda-epoch=20160418=1
multi_key_dict=2.0.3=py_0
munch=4.0.0=pypi_0
nbconvert=5.6.1=py36_0
nbformat=5.0.4=py_0
notebook=6.4.0=py36haa95532_0
numpy=1.18.1=py36h93ca92e_0
numpy-base=1.18.1=py36hc3f5095_1
numpydoc=0.9.2=py_0
oauthlib=3.1.0=py_0
olefile=0.46=py36_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1w=h2bbff1b_0
packaging=20.3=py_0
pandas=1.0.3=py36h47e9c7a_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
paramiko=2.7.1=py_0
parso=0.5.2=py_0
passlib=1.7.4=pypi_0
pathtools=0.1.2=py_1
pbr=5.6.0=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pillow=8.4.0=pypi_0
pip=20.0.2=py36_1
pluggy=0.13.1=py36_0
proj4=5.2.0=ha925a31_1
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
protobuf=3.12.3=pypi_0
psutil=5.7.0=py36he774522_0
psycopg2=2.8.4=py36h7a1dbc1_0
py=1.8.1=py_0
pyasn1=0.4.8=pyhd3eb1b0_0
pycodestyle=2.5.0=py36_0
pycparser=2.20=py_0
pycryptodome=3.9.7=pypi_0
pydocstyle=4.0.1=py_0
pyflakes=2.1.1=py36_0
pygithub=1.51=py_0
pygments=2.6.1=py_0
pyjwt=1.7.1=pypi_0
pykeepass=3.2.0=pypi_0
pylint=2.4.4=py36_0
pymongo=3.9.0=py36ha925a31_0
pynacl=1.3.0=py36h62dcd97_0
pyodbc=4.0.30=py36ha925a31_0
pyopenssl=19.1.0=py36_0
pyotp=2.7.0=pypi_0
pyparsing=2.4.6=py_0
pyproj=3.0.1=pypi_0
pyqt=5.9.2=py36h6538335_2
pyreadline=2.1=py36_1
pyrsistent=0.16.0=py36he774522_0
pysal=1.14.4.post1=py36_1
pysftp=0.2.9=py_1
pysocks=1.7.1=py36_0
pytest=5.4.1=py36_0
python=3.6.9=h5500b2f_0
python-dateutil=2.8.1=py_0
python-jenkins=1.6.0=py_0
python-jsonrpc-server=0.3.4=py_0
python-language-server=0.31.10=py36_0
python_abi=3.6=1_cp36m
pythonnet=2.3.0=py36_0
pytz=2019.3=py_0
pywin32=227=py36he774522_1
pywin32-ctypes=0.2.0=py36_1000
pywinpty=0.5.7=py36_0
pyyaml=5.3.1=py36he774522_0
pyzmq=18.1.1=py36ha925a31_0
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=0.7.0=py_0
qtconsole=4.7.3=py_0
qtpy=1.9.0=py_0
requests=2.22.0=py36_1
requests-kerberos=0.12.0=py36haa95532_0
requests-oauthlib=1.3.0=py_0
requests-toolbelt=0.9.1=pyhd3eb1b0_0
rope=0.16.0=py_0
rtree=0.9.3=py36h21ff451_0
scipy=1.5.0=pypi_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=46.1.3=py36_0
shapely=1.6.4=py36h222a598_0
sip=4.19.8=py36h6538335_0
six=1.14.0=py36_0
smmap=4.0.0=pyhd3eb1b0_0
snowballstemmer=2.0.0=py_0
soaplib=1.0.0=pypi_0
sortedcontainers=2.1.0=py36_0
sphinx=2.4.4=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.2=py36_0
spyder-kernels=1.9.0=py36_0
spyder-unittest=0.4.0=py_0
spyne=2.14.0=pyhd8ed1ab_0
sqlalchemy=1.3.16=py36he774522_0
sqlite=3.31.1=h2a8f88b_1
sqlparse=0.4.4=pypi_0
suds=1.1.2=pypi_0
tensorboard=1.14.0=py36he3c9ec2_0
tensorflow=1.14.0=eigen_py36hf4fd08c_0
tensorflow-base=1.14.0=eigen_py36hdbc3f0e_0
tensorflow-estimator=1.14.0=py_0
termcolor=1.1.0=py36haa95532_1
terminado=0.9.4=py36haa95532_0
testpath=0.4.4=py_0
tk=8.6.12=h2bbff1b_0
tornado=6.0.4=py36he774522_1
traitlets=4.3.3=py36_0
typed-ast=1.4.1=py36he774522_0
typing_extensions=4.1.1=pyh06a4308_0
ua-parser=0.16.1=pypi_0
ujson=1.35=py36hfa6e2cd_0
urllib3=1.25.8=py36_0
user-agents=2.2.0=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
waitress=2.0.0=pyhd3eb1b0_0
watchdog=0.10.2=py36_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py36_1
websocket-client=0.58.0=py36haa95532_4
werkzeug=2.0.3=pyhd3eb1b0_0
wfastcgi=3.0.0=py36ha15d459_3
wheel=0.34.2=py36_0
widgetsnbextension=3.4.2=py36_0
win_inet_pton=1.1.0=py36_0
wincertstore=0.2=py36h7fe50ca_0
winkerberos=0.7.0=py36he774522_1
winpty=0.4.3=4
wrapt=1.12.1=py36he774522_1
xerces-c=3.2.3=ha925a31_0
xmldiff=2.4=pyhd8ed1ab_0
xz=5.4.5=h8cc25b3_0
yaml=0.1.7=hc54c509_2
yapf=0.28.0=py_0
zeromq=4.3.1=h33f27b4_3
zipp=2.2.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.5.0=h19a0ad4_1
